Restored app_home spec. Removed privacy_compliance_webhooks spec
