+++ packages/app/src/cli/models/extensions/specifications/app_config_app_home.test.ts
@@ -1,0 +1,53 @@
+import spec from './app_config_app_home.js'
+import {describe, expect, test} from 'vitest'
+
+describe('app_home', () => {
+  describe('transform', () => {
+    test('should return the transformed object', () => {
+      // Given
+      const object = {
+        application_url: 'https://my-app-url.dev',
+        embedded: true,
+        app_preferences: {
+          url: 'https://my-preferences-url.dev',
+        },
+      }
+
+      const appConfigSpec = spec
+
+      // When
+      const result = appConfigSpec.transform!(object)
+
+      // Then
+      expect(result).toMatchObject({
+        app_url: 'https://my-app-url.dev',
+        embedded: true,
+        preferences_url: 'https://my-preferences-url.dev',
+      })
+    })
+  })
+
+  describe('reverseTransform', () => {
+    test('should return the reversed transformed object', () => {
+      // Given
+      const object = {
+        app_url: 'https://my-app-url.dev',
+        embedded: true,
+        preferences_url: 'https://my-preferences-url.dev',
+      }
+      const appConfigSpec = spec
+
+      // When
+      const result = appConfigSpec.reverseTransform!(object)
+
+      // Then
+      expect(result).toMatchObject({
+        application_url: 'https://my-app-url.dev',
+        embedded: true,
+        app_preferences: {
+          url: 'https://my-preferences-url.dev',
+        },
+      })
+    })
+  })
+})
